refactor: split public key filter helpers out of list.go
Keep list.go under the backend structure 1200-line gate.
